[backend] feat(service-account): review request: merge tests capability.
