+++ openaev-api/src/test/java/io/openaev/api/capabilities/CapabilityTreeBuilderTest.java
@@ -6,8 +6,12 @@ import static io.openaev.database.model.CapabilityScope.PLATFORM;
 import static io.openaev.database.model.CapabilityScope.TENANT;
 import static org.assertj.core.api.Assertions.*;
 
+import io.openaev.database.model.Capability;
 import io.openaev.database.model.CapabilityScope;
 import java.util.List;
+
+import io.openaev.service.account.Constants;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 class CapabilityTreeBuilderTest {
@@ -193,5 +197,29 @@ class CapabilityTreeBuilderTest {
       result.addAll(flattenValues(node.children()));
     }
     return result;
+  }
+
+  @Test
+  @DisplayName("Service role capabilities should contain AGENT_RUNTIME_ACCESS and ACCESS_DOCUMENTS")
+  void given_serviceRoleCapabilities_should_containExpectedCapabilities() {
+    // -- ASSERT --
+    assertThat(Constants.SERVICE_ROLE_CAPABILITIES)
+        .containsExactlyInAnyOrder(Capability.AGENT_RUNTIME_ACCESS, Capability.ACCESS_DOCUMENTS);
+  }
+
+  @Test
+  @DisplayName("Service role name and description should be defined")
+  void given_serviceRoleConstants_should_haveNameAndDescription() {
+    // -- ASSERT --
+    assertThat(Constants.SERVICE_ROLE_NAME).isEqualTo("Service integration");
+    assertThat(Constants.SERVICE_ROLE_DESCRIPTION).isNotBlank();
+  }
+
+  @Test
+  @DisplayName("Service group name and description should be defined")
+  void given_serviceGroupConstants_should_haveNameAndDescription() {
+    // -- ASSERT --
+    assertThat(Constants.SERVICE_GROUP_NAME).isEqualTo("Service integration");
+    assertThat(Constants.SERVICE_GROUP_DESCRIPTION).isNotBlank();
   }
 }
